# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Benetnasch API
  version: 1.0.0
  description: Versioned HTTP contract for the Benetnasch blog and administration apps.
servers:
  - url: /api/v1
...
paths:
...
  /auth/logout:
    post: { tags: [auth], summary: Sign out, operationId: logout, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /auth/me:
    put: { tags: [auth], summary: Update current user, operationId: updateCurrentUser, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/dashboard:
    get: { tags: [admin], summary: Get dashboard information, operationId: getAdminDashboard, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles:
    get: { tags: [admin], summary: List admin articles, operationId: listAdminArticles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save an article, operationId: saveAdminArticle, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/articles/featured:
    put: { tags: [admin], summary: Update article flags, operationId: updateArticleFlags, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/trash:
    put: { tags: [admin], summary: Move articles to or from trash, operationId: updateArticleTrash, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/articles/batch-delete:
    delete: { tags: [admin], summary: Permanently delete articles, operationId: deleteArticles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/site:
    get: { tags: [admin], summary: Get website configuration, operationId: getSite, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    put: { tags: [admin], summary: Update website configuration, operationId: updateSite, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/friend-links:
    get: { tags: [admin], summary: List friend links, operationId: listAdminFriendLinks, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save friend link, operationId: saveFriendLink, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete friend links, operationId: deleteFriendLinks, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/albums:
    get: { tags: [admin], summary: List albums, operationId: listAdminAlbums, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save album, operationId: saveAlbum, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/albums/options:
    get: { tags: [admin], summary: List album options, operationId: listAlbumOptions, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/albums/cover:
    post: { tags: [admin], summary: Upload album cover, operationId: uploadAlbumCover, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/photos/trash:
    put: { tags: [admin], summary: Update photo trash status, operationId: updatePhotoTrash, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/permissions:
    get: { tags: [admin], summary: List permissions, operationId: listPermissions, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save permission, operationId: savePermission, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/roles/resource-options:
    get: { tags: [admin], summary: List role resource options, operationId: listRoleResourceOptions, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/roles/menu-options:
    get: { tags: [admin], summary: List role menu options, operationId: listRoleMenuOptions, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/menus:
    get: { tags: [admin], summary: List menus, operationId: listMenus, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    post: { tags: [admin], summary: Save menu, operationId: saveMenu, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/me/menu:
    get: { tags: [admin], summary: Get current user menu, operationId: getCurrentUserMenu, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/users/roles:
    get: { tags: [admin], summary: List user role options, operationId: listUserRoles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    put: { tags: [admin], summary: Update user roles, operationId: updateUserRoles, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/users/areas:
    get: { tags: [admin], summary: Get user areas, operationId: listUserAreas, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/jobs/groups:
    get: { tags: [admin], summary: List job groups, operationId: listJobGroups, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
...
  /admin/logs/jobs:
    get: { tags: [admin], summary: List job logs, operationId: listJobLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete job logs, operationId: deleteJobLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/logs/jobs/clean:
    delete: { tags: [admin], summary: Clean job logs, operationId: cleanJobLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/logs/jobs/groups:
    get: { tags: [admin], summary: List job log groups, operationId: listJobLogGroups, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/logs/operations:
    get: { tags: [admin], summary: List operation logs, operationId: listOperationLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete operation logs, operationId: deleteOperationLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
  /admin/logs/exceptions:
    get: { tags: [admin], summary: List exception logs, operationId: listExceptionLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
    delete: { tags: [admin], summary: Delete exception logs, operationId: deleteExceptionLogs, security: [{ bearerAuth: [] }], responses: { '200': { $ref: '#/components/responses/Result' } } }
components:
...
  responses:
    Result:
      description: Versioned API response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data]
      properties:
        code: { type: string, example: OK }
        message: { type: string, example: 操作成功 }
        data: { nullable: true }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT